+++ rd-api-client/src/main/java/org/rundeck/client/api/model/JobFileItem.java
@@ -17,9 +17,9 @@
 package org.rundeck.client.api.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
 import org.rundeck.client.util.DataOutput;
 
-import javax.xml.bind.annotation.XmlRootElement;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -27,8 +27,8 @@ import java.util.Map;
  * @author greg
  * @since 3/1/17
  */
+@Data
 @JsonIgnoreProperties(ignoreUnknown = true)
-@XmlRootElement()
 public class JobFileItem implements DataOutput {
     private String id;
     private String user;
@@ -43,108 +43,12 @@ public class JobFileItem implements DataOutput {
     private String expirationDate;
     private Long execId;
 
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getUser() {
-        return user;
-    }
-
-    public void setUser(String user) {
-        this.user = user;
-    }
-
-    public String getOptionName() {
-        return optionName;
-    }
-
-    public void setOptionName(String optionName) {
-        this.optionName = optionName;
-    }
-
-    public String getFileState() {
-        return fileState;
-    }
-
-    public void setFileState(String fileState) {
-        this.fileState = fileState;
-    }
-
-    public String getSha() {
-        return sha;
-    }
-
-    public void setSha(String sha) {
-        this.sha = sha;
-    }
-
-    public String getJobId() {
-        return jobId;
-    }
-
-    public void setJobId(String jobId) {
-        this.jobId = jobId;
-    }
-
-    public String getDateCreated() {
-        return dateCreated;
-    }
-
     public DateInfo dateInfoDateCreated() {
         return new DateInfo(dateCreated);
     }
 
-    public void setDateCreated(String dateCreated) {
-        this.dateCreated = dateCreated;
-    }
-
-    public String getServerNodeUUID() {
-        return serverNodeUUID;
-    }
-
-    public void setServerNodeUUID(String serverNodeUUID) {
-        this.serverNodeUUID = serverNodeUUID;
-    }
-
-    public String getFileName() {
-        return fileName;
-    }
-
-    public void setFileName(String fileName) {
-        this.fileName = fileName;
-    }
-
-    public Long getSize() {
-        return size;
-    }
-
-    public void setSize(Long size) {
-        this.size = size;
-    }
-
-    public String getExpirationDate() {
-        return expirationDate;
-    }
-
     public DateInfo dateInfoExpirationDate() {
         return new DateInfo(expirationDate);
-    }
-
-    public void setExpirationDate(String expirationDate) {
-        this.expirationDate = expirationDate;
-    }
-
-    public Long getExecId() {
-        return execId;
-    }
-
-    public void setExecId(Long execId) {
-        this.execId = execId;
     }
 
     public Map<?, ?> asMap() {