+++ rd-api-client/src/test/groovy/org/rundeck/client/api/model/ImportResultSpec.groovy
@@ -4,55 +4,59 @@ import okhttp3.MediaType
 import okhttp3.ResponseBody
 import retrofit2.Converter
 import retrofit2.Retrofit
+import retrofit2.converter.jackson.JacksonConverterFactory
 import retrofit2.converter.jaxb.JaxbConverterFactory
 import spock.lang.Specification
 
 import java.lang.annotation.Annotation
 
 class ImportResultSpec extends Specification {
-    def "parse import result xml"(){
+    def "parse import result json"() {
         given:
-            def xmlText='''<result success='true' apiversion='35'>
-  <succeeded count='0' />
-  <failed count='2'>
-    <job index='1'>
-      <name>test00</name>
-      <group>group00</group>
-      <project>bongo</project>
-      <error>Cannot create a Job with UUID 17b07de4-f47d-4ab1-9f3a-8d47b33058c4: a Job already exists...</error>
-    </job>
-    <job index='2'>
-      <name>test01</name>
-      <group></group>
-      <project>bongo</project>
-      <error>...errr</error>
-    </job>
-  </failed>
-  <skipped count='2'>
-    <job index='12'>
-      <name>test02</name>
-      <group>group02</group>
-      <project>bongo</project>
-      
-      <error>...errr</error>
-    </job>
-    <job index='13'>
-      <name>test03</name>
-      <group></group>
-      <project>bongo</project>
-      <error>...errr</error>
-    </job>
-  </skipped> 
-</result>
+            def text='''{
+"succeeded": [],
+"failed": [
+    {
+        "index":1,
+        "name":"test00",
+        "group":"group00",
+        "project":"bongo",
+        "error":"Cannot create a Job with UUID 17b07de4-f47d-4ab1-9f3a-8d47b33058c4: a Job already exists..."
+    },
+    {
+        "index":2,
+        "name":"test01",
+        "group":"",
+        "project":"bongo",
+        "error":"...errr"
+    }
+],
+"skipped":[
+    {
+        "index":12,
+        "name":"test02",
+        "group":"group02",
+        "project":"bongo",
+        "error":"...errr"
+    },
+    {
+        "index":13,
+        "name":"test03",
+        "group":"",
+        "project":"bongo",
+        "error":"...errr"
+    }
+]
+}
 '''
             def retrofit = new Retrofit.Builder().baseUrl('http://test').
-                addConverterFactory(JaxbConverterFactory.create()).
+                addConverterFactory(JacksonConverterFactory.create()).
                 build()
 
         when:
             Converter<ResponseBody, ImportResult> converter = retrofit.
                 responseBodyConverter(ImportResult.class, [] as Annotation[],);
-            ImportResult result = converter.convert(ResponseBody.create(MediaType.parse('application/xml'), xmlText))
+            ImportResult result = converter.convert(ResponseBody.create(text, MediaType.parse('application/json')))
 
         then:
             result.succeeded!=null